requirements.txt
bcrypt==3.1.7
cffi==1.14.0
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
kiwisolver==1.1.0
matplotlib==3.2.1
netaddr==0.7.19
networkx==2.4
numpy==1.18.2
pandas==1.0.3
paramiko==2.7.1
pycparser==2.20
PyNaCl==1.3.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
telnetlib3==1.0.2
